user_interface/requirements.txt
# Core libraries
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
scikit-learn>=1.0.2

# Deep learning
torch==2.5.0
torchvision>=0.14.0
torchaudio>=2.5.0

# Lightning
pytorch_lightning>=2.0.0
lightning>=2.0.0

# Segmentation models wrapper
segmentation-models-pytorch>=0.3.2

# Metrics
torchmetrics>=1.0.0

# Streamlit GUI
streamlit>=1.17.0